fix: Revert exmaple
