+++ modules/react/toast/stories/examples/Basic.tsx
@@ -2,32 +2,14 @@ import React from 'react';
 
 import {Toast} from '@workday/canvas-kit-react/toast';
 import {checkIcon} from '@workday/canvas-system-icons-web';
-import {Popper} from '@workday/canvas-kit-react/popup';
-import {SecondaryButton} from '@workday/canvas-kit-react/button';
 
 export const Basic = () => {
-  const [open, setOpen] = React.useState(false);
-  const containerRef = React.useRef(null);
-
-  const handleClose = () => {
-    setOpen(false);
-  };
-
-  const handleOpen = () => {
-    setOpen(true);
-  };
   return (
-    <div ref={containerRef}>
-      <SecondaryButton onClick={handleOpen}>Show Status Toast</SecondaryButton>
-      <Popper placement="top" open={open} anchorElement={containerRef}>
-        <Toast>
-          <Toast.Icon icon={checkIcon} color="greenApple400" />
-          <Toast.Body>
-            <Toast.Message>Your workbook was successfully processed.</Toast.Message>
-          </Toast.Body>
-          <Toast.CloseIcon onClick={handleClose} />
-        </Toast>
-      </Popper>
-    </div>
+    <Toast>
+      <Toast.Icon icon={checkIcon} color="greenApple400" />
+      <Toast.Body>
+        <Toast.Message>Your workbook was successfully processed.</Toast.Message>
+      </Toast.Body>
+    </Toast>
   );
 };
